Added timer to stopwatch
Second and minute hands both now move and keep track of time once stopwatch mode is started.
There is an issue though (99.9% chance it's in Stopwatch) where I can't find a way to draw the hands directly on top of the dial. Either its off to the side doing its down thing independent to the dial or (when trying to layer it), it completely hides the dial